move misplaced tests in test_ptable_plotly into test_ptable_heatmap_plotly and new test_ptable_heatmap_splits_plotly
